requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
openpyxl==3.1.2
plotly==5.18.0
reportlab==4.0.9
python-pptx==0.6.23
pydantic==2.5.3
pydantic-settings==2.1.0
